+++ Nos/Extensions/String+Markdown.swift
@@ -22,7 +22,7 @@ extension String {
                 var link = linkDisplayName
                 if var url = URL(string: link) {
                     if url.scheme == nil,
-                        let httpsURL = URL(string: ("https://\(link)")) {
+                       let httpsURL = URL(string: ("https://\(link)")) {
                         url = httpsURL
                     }
                     link = url.absoluteString
@@ -34,7 +34,7 @@ extension String {
         return string
     }
     
-    /// Creates a new string with all URLs and any preceding whitespace removed, and returns the new string and an 
+    /// Creates a new string with all URLs and any preceding and trailing whitespace removed and removed duplicate newlines, and returns the new string and an
     /// array of all the URLs.
     func extractURLs() -> (String, [URL]) {
         var urls: [URL] = []
@@ -57,9 +57,13 @@ extension String {
             Log.error("Invalid regex pattern")
         }
         
+        mutableString.replaceOccurrences(of: "^\\s*", with: "", options: .regularExpression, range: NSRange(location: 0, length: mutableString.length))
+        mutableString.replaceOccurrences(of: "\\s*$", with: "", options: .regularExpression, range: NSRange(location: 0, length: mutableString.length))
+        mutableString.replaceOccurrences(of: "\\n{2,}", with: "\n", options: .regularExpression, range: NSRange(location: 0, length: mutableString.length))
+        
         return (mutableString as String, urls)
     }
-
+    
     func findUnformattedLinks() throws -> [URL] {
         // swiftlint:disable line_length
         let regex = "((http|https)?:\\/\\/.)?(www\\.)?[-a-zA-Z0-9@:%._\\+~#=]{2,256}\\.[a-z]{2,6}\\b([-a-zA-Z0-9@:%_\\+.~#?&//=]*)"
@@ -70,7 +74,7 @@ extension String {
         let wholeRange = NSRange(location: 0, length: self.utf16.count)
         for match in regularExpression.matches(in: self, range: wholeRange) {
             if let range = Range(match.range, in: self),
-                let url = URL(string: String(self[range])) {
+               let url = URL(string: String(self[range])) {
                 links.append(url)
             }
         }